fix bug camera, add new classes: moving rock, falling rock

+++ 05-SceneManager/QuadTree.cpp
@@ -82,7 +82,7 @@ void Quadtree::Insert(LPGAMEOBJECT entity)
     }
 
     Camera* cam = Camera::GetInstance();
-    float cWidth, cHeight;
+    int cWidth, cHeight;
     cam->getCamWidth(cWidth);
     cam->getCamHeight(cHeight);
